# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QUANTMINE Web API
  version: 0.0.0
  description: |
    阶段 0 契约草案。覆盖 FRONTEND_DESIGN_HANDOFF.md §7 全部端点，
    不含任何业务实现细节。错误响应严格遵循 §8 统一错误格式。
servers:
  - url: http://localhost:8000
    description: Local dev (FastAPI)
...
components:
...
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    ValidationFailed:
      description: 参数校验失败（422）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      required: [code, title, status]
      properties:
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_FAILED
            - CONFLICT
            - RATE_LIMITED
            - UPSTREAM_FAILURE
            - TASK_RUNNING
            - INTERNAL_ERROR
        title: { type: string }
        detail: { type: string }
        traceId: { type: string }
        status: { type: integer }
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    SeriesPoint:
      type: object
      required: [date, value]
      properties:
        date: { type: string, format: date }
        value: { type: number }
    SeriesResponse:
      type: object
      required: [baseDate, series]
      properties:
        baseDate: { type: string, format: date }
        series:
          type: array
          items:
            type: object
            required: [symbol, points]
            properties:
              symbol: { type: string }
              points:
                type: array
                items: { $ref: '#/components/schemas/SeriesPoint' }

    # —— Rebalance ——
...
paths:
...
  /api/v1/market/series:
    get:
      tags: [market]
      summary: 多 ticker 时间序列（按区间 + 频率查询）
      parameters:
        - in: query
          name: symbols
          required: true
          schema: { type: array, items: { type: string } }
        - in: query
          name: startDate
          required: true
          schema: { type: string, format: date }
        - in: query
          name: endDate
          required: true
          schema: { type: string, format: date }
        - in: query
          name: frequency
          schema: { type: string, enum: [D, W, M], default: D }
        - in: query
          name: normalize
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesResponse' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        default: { $ref: '#/components/responses/Error' }
...